users app added

+++ blog/views.py
@@ -1,8 +1,6 @@
-from urllib import request
-from django.shortcuts import render, redirect
+from django.shortcuts import render, redirect, get_object_or_404, HttpResponse
 from .forms import CommentForm, PostForm
 from blog.models import Like, Post
-from django.shortcuts import get_object_or_404
 
 # Create your views here.
 
@@ -49,6 +47,9 @@ def post_detail(request, slug):
 
 def post_update(request, slug):
     obj = Post.objects.get(slug=slug)
+    if request.user.id != obj.author.id:
+        # return HttpResponse("You are not authorized!!")
+        return redirect("blog:list")
     form = PostForm(request.POST or None, request.FILES or None, instance=obj)
     if form.is_valid():
         form.save()
@@ -62,6 +63,9 @@ def post_update(request, slug):
 
 def post_delete(request, slug):
     obj = get_object_or_404(Post, slug=slug)
+    if request.user.id != obj.author.id:
+        # return HttpResponse("You are not authorized!!")
+        return redirect("blog:list")
     if request.method == 'POST':
         obj.delete()
         return redirect('blog:list')
